requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-multipart
openai
requests
python-dotenv
